dataflow/requirements.txt
google-cloud-vision==0.38.0
apache-beam[gcp]
grpcio
oauth2client==3.0.0
httplib2==0.15.0
